remove specification of KATEX_PATH
This is for the pelican-katex plugin. Disabling the setting of
KATEX_PATH will mean that pelican-katex uses the version of KaTeX
deployed alongside pelican-katex itself.

Unfortunately, NodeJS does not (without extra effort) look for modules
installed outside the current working directory when attempting to load
a module using require(), so, even if KaTeX is installed globally, there
would be extra setup in order for pelican-katex to be able to use the
developer-installed KaTeX.

See the following Stack Overflow answer:
    https://stackoverflow.com/a/48140439

The paradigm in NodeJS seems to be "install everything locally".